refactor: access control
